requirements.txt
snowflake-snowpark-python
pydantic
pyyaml
pytest
ruff
